+++ tests/integration/clearingHouse/positionTest.ts
@@ -1,7 +1,7 @@
-import { web3 } from "hardhat"
 import { expectEvent, expectRevert } from "@openzeppelin/test-helpers"
 import { default as BN } from "bn.js"
 import { use } from "chai"
+import { web3 } from "hardhat"
 import {
     AmmFakeInstance,
     ClearingHouseFakeInstance,
@@ -846,7 +846,7 @@ describe("ClearingHouse - open/close position Test", () => {
             )
         })
 
-        it("alice take profit from bob's unrealized under-collateral position, then bob close", async () => {
+        it("alice takes profit from bob's position, putting his position underwater, then bob closes", async () => {
             // alice opens short position
             await approve(alice, clearingHouse.address, 20)
             await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(20), toDecimal(10), toDecimal(0), {
@@ -885,7 +885,7 @@ describe("ClearingHouse - open/close position Test", () => {
             expect(await quoteToken.balanceOf(clearingHouse.address)).eq(0)
         })
 
-        it("alice take profit from bob's unrealized under-collateral position, then bob got liquidate", async () => {
+        it("alice takes profit from bob's position, putting his position underwater, then bob gets liquidated", async () => {
             // alice opens short position
             await approve(alice, clearingHouse.address, 20)
             await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(20), toDecimal(10), toDecimal(0), {
@@ -925,7 +925,8 @@ describe("ClearingHouse - open/close position Test", () => {
             expect(await quoteToken.balanceOf(carol)).eq("14705882")
         })
 
-        it("alice's position got liquidated but still has enough margin left for paying liquidation fee", async () => {
+        // the test for pointing out the calculation of margin ratio should be based on positionNotional instead of openNotional
+        it("alice's position has enough margin left, thus won't get liquidated", async () => {
             // alice opens long position
             await approve(alice, clearingHouse.address, 300)
             await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(300), toDecimal(2), toDecimal(0), {
@@ -938,23 +939,15 @@ describe("ClearingHouse - open/close position Test", () => {
                 from: bob,
             })
 
-            // alice's margin ratio = (margin + unrealizedPnl) / openNotional = (300 + (-278.77)) / 600 = 3.53%
-            const receipt = await clearingHouse.liquidate(amm.address, alice, { from: carol })
-
-            // liquidationFee = 321.23 * 5% = 16.06
-            // remainMargin = margin + unrealizedPnl = 300 + (-278.77) = 21.23
-            // Since 21.23 - 16.06 >= 0, both badDebts = 0
-            // Trader total PnL = -278.77 - 21.23 = -300 since she lost her remaining margin to the insurance fund
-
-            expectEvent(receipt, "PositionChanged", {
-                realizedPnl: "-278761061946902654868",
-                badDebt: "0",
-                liquidationPenalty: "21238938053097345132",
-            })
-            expectEvent(receipt, "PositionLiquidated", {
-                liquidationFee: "16061946902654867256",
-                badDebt: "0",
-            })
+            // unrealizedPnl: -278.77
+            // positionNotional: 600 - 278.77 = 321.23
+            // remainMargin: 300 - 278.77 = 21.23
+            // liquidationFee: 321.23 * 5% = 16.06
+            // margin ratio: = (margin + unrealizedPnl) / positionNotional = 21.23 / 321.23 = 6.608971765%
+            await expectRevert(
+                clearingHouse.liquidate(amm.address, alice, { from: carol }),
+                "Margin ratio not meet criteria",
+            )
         })
 
         it("alice's position got liquidated and not enough margin left for paying liquidation fee", async () => {
@@ -991,7 +984,101 @@ describe("ClearingHouse - open/close position Test", () => {
             })
         })
 
-        it("force error, can NOT open a long/short position when position(long) is under collateral", async () => {
+        it("can open position of the same side even though position(long) is underwater, as long as the margin ratio will be over maintenance ratio after the action", async () => {
+            await approve(alice, clearingHouse.address, 2000)
+            await approve(bob, clearingHouse.address, 2000)
+
+            // alice gets 20 position
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(25), toDecimal(10), toDecimal(0), {
+                from: alice,
+            })
+            // AMM after 1250 : 80
+
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: bob,
+            })
+            // AMM after 1000 : 100
+
+            /**
+             * position size = 20
+             * margin = 25
+             * positionNotional = 166.67
+             * openNotional = 250
+             * unrealizedPnl = 166.67 - 250 = -83.33
+             * marginRatio = (25 + -83.33) / 250 = -23%
+             */
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(100), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+
+            /*
+             * AMM after 1100 : 90.90909
+             * positionNotional = 166.67 + 100 = 266.67
+             * position size = 20 + 9.09 = 29.09
+             * margin = 25 + 100
+             */
+            expect((await clearingHouse.getPosition(amm.address, alice)).margin).to.eq("125000000000000000000")
+            expect((await clearingHouse.getPosition(amm.address, alice)).size).to.eq("29090909090909090909")
+            expect(
+                (
+                    await clearingHouse.getPositionNotionalAndUnrealizedPnl(
+                        amm.address,
+                        alice,
+                        PnlCalcOption.SPOT_PRICE,
+                    )
+                )[0],
+            ).to.eq("266666666666666666665")
+        })
+
+        it("can open reverse position even though position(long) is underwater, as long as the margin ratio will be over maintenance ratio after the action", async () => {
+            await approve(alice, clearingHouse.address, 2000)
+            await approve(bob, clearingHouse.address, 2000)
+
+            // alice gets about 20 position
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(88), toDecimal(2.841), toDecimal(0), {
+                from: alice,
+            })
+            expect((await clearingHouse.getPosition(amm.address, alice)).margin).to.eq(toFullDigit(88))
+            // AMM after 1250 : 80
+
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: bob,
+            })
+            // AMM after 1000 : 100
+
+            /**
+             * position size = 20
+             * margin = 88
+             * positionNotional = 166.67
+             * openNotional ~= 250
+             * unrealizedPnl = 166.67 - 250 = -83.33
+             * marginRatio = (88 + -83.33) / 250 = 1.868%
+             */
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(150), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+
+            /**
+             * AMM after 850 : 117.64705882
+             * positionNotional = 166.67 - 150 = 16.67
+             * position size = 20 - 17.64705882 = 2.35
+             * realizedPnl = -83.33 * (20 - 2.35) / 20 = -73.538725
+             * margin = 88 -73.538725 ~= 14.4
+             */
+            expect((await clearingHouse.getPosition(amm.address, alice)).size).to.eq("2353760435608511085")
+            expect((await clearingHouse.getPosition(amm.address, alice)).margin).to.eq("14471986140208919751")
+            expect(
+                (
+                    await clearingHouse.getPositionNotionalAndUnrealizedPnl(
+                        amm.address,
+                        alice,
+                        PnlCalcOption.SPOT_PRICE,
+                    )
+                )[0],
+            ).to.eq("16672666683555438214")
+        })
+
+        it("force error, cannot open position if position(long) is underwater and will still be after the action", async () => {
             // deposit to 2000
             await approve(alice, clearingHouse.address, 2000)
             await approve(bob, clearingHouse.address, 2000)
@@ -1027,7 +1114,7 @@ describe("ClearingHouse - open/close position Test", () => {
             )
         })
 
-        it("force error, can NOT open a long/short position when position(short) is under collateral", async () => {
+        it("force error, cannot open position if position(short) is underwater and will still be after the action", async () => {
             // deposit to 2000
             await approve(alice, clearingHouse.address, 2000)
             await approve(bob, clearingHouse.address, 2000)